+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Chaned
-- Fixed string parsing by correctly handling "string continue" sequences
+- Fixed (byte) string literal parsing by:
+    - Correctly handling "string continue" sequences
+    - Correctly converting `\n\r` into `\n`
 
 
 ## [0.2.1] - 2021-06-04