+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 
 
 ## [Unreleased]
+
+
+## [0.2.2] - 2021-06-09
 ### Changed
 - Fixed (byte) string literal parsing by:
     - Correctly handling "string continue" sequences
     - Correctly converting `\n\r` into `\n`
-
 
 ## [0.2.1] - 2021-06-04
 ### Changed
@@ -46,7 +48,8 @@ All notable changes to this project will be documented in this file.
 - Everything
 
 
-[Unreleased]: https://github.com/LukasKalbertodt/litrs/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/LukasKalbertodt/litrs/compare/v0.2.2...HEAD
+[0.2.2]: https://github.com/LukasKalbertodt/litrs/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/LukasKalbertodt/litrs/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/LukasKalbertodt/litrs/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/LukasKalbertodt/litrs/compare/v0.1.0...v0.1.1